Move IAuthor to models directory
